taxonomy: use data space analytics API for loading
