+++ .changeset/quick-panthers-collect.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-taxonomy': minor
+---
+
+Make use of data space analytics API to render the data spaces for better performance. Also improved loading mechanism to avoid memory leaks.
