+++ packages/legend-extension-dsl-data-space/src/index.ts
@@ -28,3 +28,5 @@ export {
 export { DataSpaceViewerState } from './stores/DataSpaceViewerState.js';
 
 export * from './graphManager/DSLDataSpace_GraphManagerHelper.js';
+export { getDSLDataSpaceGraphManagerExtension } from './graphManager/protocol/DSLDataSpace_PureGraphManagerExtension.js';
+export { DataSpaceAnalysisResult } from './graphManager/action/analytics/DataSpaceAnalysis.js';
