+++ packages/legend-taxonomy/src/components/StandaloneDataSpaceViewer.tsx
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import { useEffect } from 'react';
+import { createContext, useContext, useEffect } from 'react';
 import { useApplicationStore } from '@finos/legend-application';
-import { observer } from 'mobx-react-lite';
+import { observer, useLocalObservable } from 'mobx-react-lite';
 import { useParams } from 'react-router';
-import type { LegendTaxonomyStandaloneDataSpaceViewerParams } from '../stores/LegendTaxonomyRouter.js';
+import type { LegendTaxonomyStandaloneDataSpaceViewerPathParams } from '../stores/LegendTaxonomyRouter.js';
 import { useLegendTaxonomyStore } from './LegendTaxonomyStoreProvider.js';
 import { flowResult } from 'mobx';
 import {
@@ -28,77 +28,114 @@ import {
   TimesCircleIcon,
 } from '@finos/legend-art';
 import { DataSpaceViewer } from '@finos/legend-extension-dsl-data-space';
+import { StandaloneDataSpaceViewerStore } from '../stores/StandaloneDataSpaceViewerStore.js';
+import type { LegendTaxonomyConfig } from '../application/LegendTaxonomyConfig.js';
+import { useDepotServerClient } from '@finos/legend-server-depot';
+import { guaranteeNonNullable } from '@finos/legend-shared';
 
-export const StandaloneDataSpaceViewer = observer(() => {
-  const params = useParams<LegendTaxonomyStandaloneDataSpaceViewerParams>();
-  const applicationStore = useApplicationStore();
-  const taxonomyStore = useLegendTaxonomyStore();
+const StandaloneDataSpaceViewerStoreContext = createContext<
+  StandaloneDataSpaceViewerStore | undefined
+>(undefined);
 
-  useEffect(() => {
-    flowResult(taxonomyStore.initializeStandaloneDataSpaceViewer(params)).catch(
-      applicationStore.alertUnhandledError,
+const StandaloneDataSpaceViewerStoreProvider: React.FC<{
+  children: React.ReactNode;
+}> = ({ children }) => {
+  const applicationStore = useApplicationStore<LegendTaxonomyConfig>();
+  const depotServerClient = useDepotServerClient();
+  const baseStore = useLegendTaxonomyStore();
+  const store = useLocalObservable(
+    () =>
+      new StandaloneDataSpaceViewerStore(
+        applicationStore,
+        depotServerClient,
+        baseStore.pluginManager,
+      ),
+  );
+  return (
+    <StandaloneDataSpaceViewerStoreContext.Provider value={store}>
+      {children}
+    </StandaloneDataSpaceViewerStoreContext.Provider>
+  );
+};
+
+export const withStandaloneDataSpaceViewerStore = (
+  WrappedComponent: React.FC,
+): React.FC =>
+  function WithStandaloneDataSpaceViewerStore() {
+    return (
+      <StandaloneDataSpaceViewerStoreProvider>
+        <WrappedComponent />
+      </StandaloneDataSpaceViewerStoreProvider>
     );
-  }, [applicationStore, params, taxonomyStore]);
-  const queryDataSpace = (): void => {
-    if (taxonomyStore.standaloneDataSpaceViewerState) {
-      taxonomyStore.queryUsingDataSpace(
-        taxonomyStore.standaloneDataSpaceViewerState,
-      );
-    }
   };
 
-  return (
-    <div className="panel standalone-data-space-viewer">
-      <PanelLoadingIndicator
-        isLoading={
-          taxonomyStore.initStandaloneDataSpaceViewerState.isInProgress
-        }
-      />
-      <div className="panel__header standalone-data-space-viewer__header">
-        <div className="panel__header__title"></div>
-        <div className="panel__header__actions standalone-data-space-viewer__header__actions">
-          <button
-            className="standalone-data-space-viewer__header__action btn--dark"
-            onClick={queryDataSpace}
-            disabled={!taxonomyStore.standaloneDataSpaceViewerState}
-            tabIndex={-1}
-            title="Query data space..."
-          >
-            <ArrowRightIcon />
-          </button>
+const useStandaloneDataSpaceViewerStore = (): StandaloneDataSpaceViewerStore =>
+  guaranteeNonNullable(
+    useContext(StandaloneDataSpaceViewerStoreContext),
+    `Can't find standalone data space viewer store in context`,
+  );
+
+export const StandaloneDataSpaceViewer = withStandaloneDataSpaceViewerStore(
+  observer(() => {
+    const params =
+      useParams<LegendTaxonomyStandaloneDataSpaceViewerPathParams>();
+    const applicationStore = useApplicationStore();
+    const viewerStore = useStandaloneDataSpaceViewerStore();
+
+    const queryDataSpace = (): void => viewerStore.queryDataSpace(undefined);
+
+    useEffect(() => {
+      flowResult(viewerStore.initialize(params)).catch(
+        applicationStore.alertUnhandledError,
+      );
+    }, [applicationStore, params, viewerStore]);
+
+    return (
+      <div className="panel standalone-data-space-viewer">
+        <div className="panel__header standalone-data-space-viewer__header">
+          <div className="panel__header__title"></div>
+          <div className="panel__header__actions standalone-data-space-viewer__header__actions">
+            <button
+              className="standalone-data-space-viewer__header__action btn--dark"
+              onClick={queryDataSpace}
+              disabled={!viewerStore.initState.hasSucceeded}
+              tabIndex={-1}
+              title="Query data space..."
+            >
+              <ArrowRightIcon />
+            </button>
+          </div>
+        </div>
+        <div className="panel__content standalone-data-space-viewer__content">
+          <PanelLoadingIndicator
+            isLoading={viewerStore.initState.isInProgress}
+          />
+          {viewerStore.viewerState && (
+            <DataSpaceViewer dataSpaceViewerState={viewerStore.viewerState} />
+          )}
+          {!viewerStore.viewerState && (
+            <>
+              {viewerStore.initState.isInProgress && (
+                <BlankPanelContent>
+                  {viewerStore.initState.message}
+                </BlankPanelContent>
+              )}
+              {viewerStore.initState.hasFailed && (
+                <BlankPanelContent>
+                  <div className="query-setup__data-space__view--failed">
+                    <div className="query-setup__data-space__view--failed__icon">
+                      <TimesCircleIcon />
+                    </div>
+                    <div className="query-setup__data-space__view--failed__text">
+                      Can&apos;t load data space
+                    </div>
+                  </div>
+                </BlankPanelContent>
+              )}
+            </>
+          )}
         </div>
       </div>
-      <div className="panel__content standalone-data-space-viewer__content">
-        {taxonomyStore.standaloneDataSpaceViewerState && (
-          <DataSpaceViewer
-            dataSpaceViewerState={taxonomyStore.standaloneDataSpaceViewerState}
-          />
-        )}
-        {!taxonomyStore.standaloneDataSpaceViewerState &&
-          taxonomyStore.initStandaloneDataSpaceViewerState.isInProgress && (
-            <BlankPanelContent>
-              {taxonomyStore.initStandaloneDataSpaceViewerState.message ??
-                taxonomyStore.graphManagerState.systemBuildState.message ??
-                taxonomyStore.graphManagerState.dependenciesBuildState
-                  .message ??
-                taxonomyStore.graphManagerState.generationsBuildState.message ??
-                taxonomyStore.graphManagerState.graphBuildState.message}
-            </BlankPanelContent>
-          )}
-        {!taxonomyStore.standaloneDataSpaceViewerState &&
-          taxonomyStore.initStandaloneDataSpaceViewerState.hasFailed && (
-            <BlankPanelContent>
-              <div className="standalone-data-space-viewer__content__placeholder--failed">
-                <div className="standalone-data-space-viewer__content__placeholder--failed__icon">
-                  <TimesCircleIcon />
-                </div>
-                <div className="standalone-data-space-viewer__content__placeholder--failed__text">
-                  Can&apos;t load data space
-                </div>
-              </div>
-            </BlankPanelContent>
-          )}
-      </div>
-    </div>
-  );
-});
+    );
+  }),
+);
