+++ packages/legend-taxonomy/style/index.scss
@@ -17,4 +17,4 @@
 @forward 'deps';
 @forward 'standalone-data-space-viewer';
 @forward 'taxonomy-node-viewer';
-@forward 'taxonomy-viewer';
+@forward 'taxonomy-explorer';
